+++ ReverseAD/include/reversead/algorithm/trivial_deriv.hpp
@@ -24,7 +24,7 @@ class TrivialDeriv {
   }
 
   TrivialDeriv(char* buf) {
-    int buf_size = 0;
+    size_t buf_size = 0;
     adjoint_vals.reset(new type_adjoint(buf));
     buf_size += adjoint_vals->byte_size();
     hessian_vals.reset(new type_hessian(&buf[buf_size]));
@@ -32,8 +32,8 @@ class TrivialDeriv {
     third_vals.reset(new type_third(&buf[buf_size]));
   }
 
-  int byte_size() const {
-    int ret = adjoint_vals->byte_size();
+  size_t byte_size() const {
+    size_t ret = adjoint_vals->byte_size();
     if (hessian_vals) {
       ret += hessian_vals->byte_size();
     }
@@ -50,7 +50,7 @@ class TrivialDeriv {
   }
 
   void write_to_byte(char* const buf) const {
-    int buf_size = 0;
+    size_t buf_size = 0;
     adjoint_vals->write_to_byte(buf);
     buf_size = adjoint_vals->byte_size();
     hessian_vals->write_to_byte(&buf[buf_size]);
