+++ ReverseAD/include/reversead/algorithm/trivial_third.hpp
@@ -48,8 +48,8 @@ class TrivialThird {
 
   // serializable
   TrivialThird(char* buf);
-  int get_size() const;
-  int byte_size() const;
+  size_t get_size() const;
+  size_t byte_size() const;
   void write_to_byte(char*) const;
   void debug() const;
 
@@ -111,8 +111,8 @@ TrivialHessian<LocType, Base> TrivialThird<LocType, Base>::get_and_erase(
 }
 
 template <typename LocType, typename Base>
-int TrivialThird<LocType, Base>::get_size() const {
-  int size_count = 0;
+size_t TrivialThird<LocType, Base>::get_size() const {
+  size_t size_count = 0;
   typename std::map<LocType, TrivialHessian<LocType, Base> >::const_iterator t_iter;
   t_iter = _data.begin();
   while (t_iter != _data.end()) {
@@ -136,16 +136,16 @@ void TrivialThird<LocType, Base>::debug() const {
 }
 
 template <typename LocType, typename Base>
-int TrivialThird<LocType, Base>::byte_size() const {
-  return get_size() * (sizeof(double) + sizeof(LocType) * 3) + sizeof(int);
+size_t TrivialThird<LocType, Base>::byte_size() const {
+  return get_size() * (sizeof(double) + sizeof(LocType) * 3) + sizeof(size_t);
 }
 
 template <typename LocType, typename Base>
 void TrivialThird<LocType, Base>::write_to_byte(char* buf) const {
 
   char* p = buf;
-  *((int*)p) = get_size();
-  p += sizeof(int);
+  *((size_t*)p) = get_size();
+  p += sizeof(size_t);
 
   typename TrivialThird<LocType, Base>::enumerator d_enum =
       this->get_enumerator();
@@ -168,11 +168,11 @@ template <typename LocType, typename Base>
 TrivialThird<LocType, Base>::TrivialThird(char* buf) {
   _data.clear();
   char* p = buf;
-  int size = 0;
+  size_t size = 0;
   LocType x, y, z;
   Base w;
-  size = *((int*)p);
-  p += sizeof(int);
+  size = *((size_t*)p);
+  p += sizeof(size_t);
   for(int i = 0; i < size; i++) {
     x = *((LocType*)p);
     p += sizeof(LocType);
