+++ ReverseAD/include/reversead/driver/forward_over_reverse.hpp
@@ -10,31 +10,31 @@
 namespace ReverseAD {
 
 void forward_over_reverse(std::shared_ptr<TrivialTrace<double>> trace,
-                          int ind_num,
-                          int dep_num,
+                          size_t ind_num,
+                          size_t dep_num,
                           double* ind_init_value,
                           double* adjoint_init_value,
                           double** hessian_vector);
 
 void forward_over_second(std::shared_ptr<TrivialTrace<double>> trace,
-                         int ind_num,
-                         int dep_num,
+                         size_t ind_num,
+                         size_t dep_num,
                          double* ind_init_value,
                          double* adjoint_init_value,
                          double*** third_vector);
 
 void forward_over_third(std::shared_ptr<TrivialTrace<double>> trace,
-                         int ind_num,
-                         int dep_num,
-                         double* ind_init_value,
-                         double* adjoint_init_value,
-                         double**** fourth_vector);
+                        size_t ind_num,
+                        size_t dep_num,
+                        double* ind_init_value,
+                        double* adjoint_init_value,
+                        double**** fourth_vector);
 
-std::shared_ptr<DerivativeTensor<int, double>> directional_reverse(
+std::shared_ptr<DerivativeTensor<size_t, double>> directional_reverse(
     std::shared_ptr<TrivialTrace<double>> trace,
-    int t_order,
-    int ind_num,
-    int dep_num,
+    size_t t_order,
+    size_t ind_num,
+    size_t dep_num,
     double* ind_init_value,
     double* adjoint_init_value);
